+++ NXNavigationExtension/Core/NXNavigationBar.m
@@ -21,9 +21,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+#import <objc/runtime.h>
+
 #import "NXNavigationBar.h"
 
-static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSUhEUgAAABgAAAAoCAMAAADT08pnAAAAhFBMVEUAAAAAg/8Ae/8AfP8Aev8Aev8Aev8Aev8Aev8Aev8Ae/8Aev8Aev8Ae/8AfP8Ae/8Aff8Afv8Ag/8Ahv8AiP8AjP8Aev8Aev8Ae/8Aev8Aev8Ae/8Ae/8Aff8Ae/8Af/8Afv8Aev8Aev8Ae/8Aev8Ae/8Ae/8Aev8Ae/8AfP8Aff8Aev+lPeVOAAAAK3RSTlMABv1G+fXr4dvW0cvAtTw4JSAWEg4K8u/mxrp1XEExKxuvpKB5b2ljUk1WM3boWAAAAMVJREFUKM9109cWgjAMBuAqigwZboaAIg7s+7+fI+3JaUL+yy8XhQzFMxviypvyq/4m5Z7oX27UD5X+p6e+1hDyyN56Tbw0fnd9Fxt/EI+A50/Xt9bPxMNpvwTgi5frG+sj8RX40nM9l7wAPzaud9ZT4j74KSP9D8H9VrnJjHdKKLRslJF5glXyAh+nv4efKzSkYZUAWyI0fWSVEMchDPCNxkdOlweXRF4reRHF1ZWWfZDOQ3vCQelePDXhOOcpAqYuk0Z9AKe5MI4L1d4yAAAAAElFTkSuQmCC";
+// <@2x.png, 默认颜色：tintColor = [UIColor systemBlueColor]
+static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSUhEUgAAABgAAAAoCAMAAADT08pnAAAAflBMVEUAAAAAkP8Aev8Ae/8Aev8Ae/8AkP8Aev8Aev8Ae/8Ae/8AfP8Afv8Afv8Af/8Aev8Aev8Ae/8Ae/8Ae/8Ae/8Aev8Aev8Ae/8Ae/8Aev8AfP8Afv8Ae/8Ae/8Ae/8Aev8Ae/8Ae/8Aev8Ae/8AfP8Aff8Ae/8AgP8AhP8Aev+USLvdAAAAKXRSTlMAAvT6910F6uFHPDcsIh3x7ubcs62lkItoYU8nzp+alYWAeXRvQDEQDnA4PYMAAAC0SURBVCjPhdNJFoIwEARQjEScEBFRcB5Qyf0vKHZ2Vd3PWv4skvSQKPm0eaLFpyF4xdsw5MTehF8q8k58PEd/Rj+gP8TTGfrd8Ju4e6Hvo9O9O/FiATy6iq/JL9Hf6GfxbIle6z7ZipdH3afkVfRcf/9mRX107JIeDqBEcgU9yjyp4Rv48RI+zqXi4mZQXG6H3UD+TwEtpyH5P1Y8iProUjrzpIE14MWxV43jXVxOTu+HRn8B1TglOoU3GxgAAAAASUVORK5CYII=";
 
 
 @implementation NXNavigationBarAppearance
@@ -90,17 +93,22 @@ static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSU
 
 @end
 
+
 @interface NXNavigationBar ()
 
 @property (nonatomic, strong) UIColor *originalBackgroundColor;
 @property (nonatomic, assign) CGRect originalNavigationBarFrame;
-@property (nonatomic, assign) UIEdgeInsets containerViewEdgeInsets;
 @property (nonatomic, assign) BOOL edgesForExtendedLayoutEnabled;
 @property (nonatomic, assign) BOOL blurEffectEnabled;
 
+@property (nonatomic, copy, class) NXNavigationBarAppearance * _Nullable(^navigationControllerAppearanceBlock)(__kindof UINavigationController *);
+
 @end
 
+
 @implementation NXNavigationBar
+
+@dynamic navigationControllerAppearanceBlock;
 
 - (instancetype)initWithFrame:(CGRect)frame {
     if (self = [super initWithFrame:frame]) {
@@ -114,9 +122,9 @@ static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSU
         _backgroundImageView.contentMode = UIViewContentModeScaleAspectFill;
         _backgroundImageView.clipsToBounds = YES;
         
-        _containerView = [[UIView alloc] init];
-        _containerView.backgroundColor = [UIColor clearColor];
-        _containerViewEdgeInsets = UIEdgeInsetsMake(0, 8, 0, 8);
+        _contentView = [[UIView alloc] init];
+        _contentView.backgroundColor = [UIColor clearColor];
+        _contentViewEdgeInsets = UIEdgeInsetsMake(0, 8, 0, 8);
         _edgesForExtendedLayoutEnabled = NO;
         
         UIBlurEffect *effect;
@@ -126,13 +134,13 @@ static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSU
             effect = [UIBlurEffect effectWithStyle:UIBlurEffectStyleExtraLight];
         }
         
-        _visualEffectView = [[UIVisualEffectView alloc] initWithEffect:effect];
-        _visualEffectView.hidden = YES;
+        _backgroundEffectView = [[UIVisualEffectView alloc] initWithEffect:effect];
+        _backgroundEffectView.hidden = YES;
         _blurEffectEnabled = NO;
         _backgroundImageView.image = [NXNavigationBarAppearance standardAppearance].backgorundImage;
         
         [self addSubview:self.backgroundImageView];
-        [self addSubview:self.visualEffectView];
+        [self addSubview:self.backgroundEffectView];
         [self addSubview:self.shadowImageView];
     }
     return self;
@@ -151,6 +159,14 @@ static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSU
 
 #pragma mark - Private
 
++ (NXNavigationBarAppearance * _Nullable (^)(__kindof UINavigationController *))navigationControllerAppearanceBlock {
+    return objc_getAssociatedObject(self, _cmd);
+}
+
++ (void)setNavigationControllerAppearanceBlock:(NXNavigationBarAppearance * _Nullable (^)(__kindof UINavigationController *))navigationControllerAppearanceBlock {
+    objc_setAssociatedObject(self, @selector(navigationControllerAppearanceBlock), navigationControllerAppearanceBlock, OBJC_ASSOCIATION_COPY_NONATOMIC);
+}
+
 - (void)setBlurEffectEnabled:(BOOL)blurEffectEnabled {
     _blurEffectEnabled = blurEffectEnabled;
     [self setBackgroundColor:self.originalBackgroundColor];
@@ -158,20 +174,20 @@ static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSU
 
 - (void)updateNavigationBarContentFrameCallSuper:(BOOL)callSuper {
     CGRect navigationBarFrame = CGRectMake(0, 0, CGRectGetWidth(_originalNavigationBarFrame), CGRectGetMaxY(_originalNavigationBarFrame));
-    self.visualEffectView.frame = navigationBarFrame;
+    self.backgroundEffectView.frame = navigationBarFrame;
     self.backgroundImageView.frame = navigationBarFrame;
     
-    CGRect containerViewFrame = CGRectMake(0, CGRectGetMinY(_originalNavigationBarFrame), CGRectGetWidth(_originalNavigationBarFrame), CGRectGetHeight(_originalNavigationBarFrame));
-    self.containerView.frame = UIEdgeInsetsInsetRect(containerViewFrame, _containerViewEdgeInsets);
+    CGRect contentViewFrame = CGRectMake(0, CGRectGetMinY(_originalNavigationBarFrame), CGRectGetWidth(_originalNavigationBarFrame), CGRectGetHeight(_originalNavigationBarFrame));
+    self.contentView.frame = UIEdgeInsetsInsetRect(contentViewFrame, _contentViewEdgeInsets);
     
     CGFloat shadowImageViewHeight = 1.0 / UIScreen.mainScreen.scale;
     self.shadowImageView.frame = CGRectMake(0, CGRectGetMaxY(_originalNavigationBarFrame) - shadowImageViewHeight, CGRectGetWidth(navigationBarFrame), shadowImageViewHeight);
     
-    // 放在所有的 View 前面，防止 containerView 被遮挡
-    if (self.superview && self.superview != self.containerView && self.superview != self.containerView.superview) {
-        [self.superview addSubview:self.containerView];
+    // 放在所有的 View 前面，防止 contentView 被遮挡
+    if (self.superview && self.superview != self.contentView && self.superview != self.contentView.superview) {
+        [self.superview addSubview:self.contentView];
     }
-    [self.superview bringSubviewToFront:self.containerView];
+    [self.superview bringSubviewToFront:self.contentView];
     
     // 重新设置 NavigationBar frame
     if (callSuper) {
@@ -180,68 +196,38 @@ static NSString *NXNavigationBarAppearanceNackImageBase64 = @"iVBORw0KGgoAAAANSU
     }
 }
 
-+ (NSMutableDictionary<NSString *, NXNavigationBarAppearance *> *)appearanceInfo {
-    static NSMutableDictionary<NSString *, NXNavigationBarAppearance *> *appearanceInfo = nil;
-    static dispatch_once_t onceToken;
-    dispatch_once(&onceToken, ^{
-        appearanceInfo = [NSMutableDictionary dictionary];
-    });
-    return appearanceInfo;
-}
-
 #pragma mark - Public
-
-- (void)enableBlurEffect:(BOOL)enabled {
-    self.blurEffectEnabled = enabled;
-}
 
 - (void)setBackgroundColor:(UIColor *)backgroundColor {
     self.originalBackgroundColor = backgroundColor;
     if (self.blurEffectEnabled) {
         self.backgroundImageView.hidden = YES;
-        self.visualEffectView.hidden = NO;
-        self.visualEffectView.contentView.backgroundColor = backgroundColor;
+        self.backgroundEffectView.hidden = NO;
+        self.backgroundEffectView.contentView.backgroundColor = backgroundColor;
         
         [super setBackgroundColor:[UIColor clearColor]];
     } else {
         self.backgroundImageView.hidden = NO;
-        self.visualEffectView.hidden = YES;
+        self.backgroundEffectView.hidden = YES;
         
         [super setBackgroundColor:backgroundColor];
     }
 }
 
-- (void)addContainerViewSubview:(UIView *)subview {
-    if (subview != self.containerView) {
-        [self.containerView addSubview:subview];
-    }
-}
-
-- (void)setContainerViewEdgeInsets:(UIEdgeInsets)edgeInsets {
-    _containerViewEdgeInsets = edgeInsets;
+- (void)setContentViewEdgeInsets:(UIEdgeInsets)contentViewEdgeInsets {
+    _contentViewEdgeInsets = contentViewEdgeInsets;
     [self updateNavigationBarContentFrameCallSuper:NO];
 }
 
-+ (NXNavigationBarAppearance *)standardAppearanceForNavigationControllerClass:(Class)aClass {
-    return [self appearanceFromRegisterNavigationControllerClass:aClass];
-}
-
-+ (void)registerStandardAppearanceForNavigationControllerClass:(Class)aClass {
-    [self registerNavigationControllerClass:aClass forAppearance:[NXNavigationBarAppearance standardAppearance]];
-}
-
-+ (NXNavigationBarAppearance *)appearanceFromRegisterNavigationControllerClass:(Class)aClass {
-    if (aClass) {
-        return [NXNavigationBar appearanceInfo][NSStringFromClass(aClass)];
++ (NXNavigationBarAppearance *)appearanceInNavigationController:(__kindof UINavigationController *)navigationController {
+    if (self.navigationControllerAppearanceBlock) {
+        return self.navigationControllerAppearanceBlock(navigationController);
     }
     return nil;
 }
 
-+ (void)registerNavigationControllerClass:(Class)aClass forAppearance:(NXNavigationBarAppearance *)appearance {
-    NSAssert(aClass != nil, @"参数不能为空！");
-    if (!aClass) return;
-    
-    [NXNavigationBar appearanceInfo][NSStringFromClass(aClass)] = appearance ?: [NXNavigationBarAppearance standardAppearance];
++ (void)setAppearanceForNavigationControllerUsingBlock:(NXNavigationBarAppearance * _Nullable (^)(__kindof UINavigationController * _Nonnull))block {
+    self.navigationControllerAppearanceBlock = block;
 }
 
 @end